refactor: dashboard components are seperated in different files
